+++ CMakeLists.txt
@@ -29,7 +29,7 @@ option(WITH_MYSQL "enable MySQL database support" ${MYSQL_FOUND})
 option(WITH_PGSQL "enable PostgreSQL database support" ${PostgreSQL_FOUND})
 option(WITH_SQLITE3 "enable SQLite database support" ${SQLITE3_FOUND})
 option(WITH_V4L2 "enable Video 4 Linux (2) webcam support" ON)
-option(WITH_PWCBSD "enable PWC webcam support (BSD only)" OFF)
+option(WITH_BKTR "enable BKTR support (BSD only)" OFF)
 
 set(HAVE_FFMPEG ${WITH_FFMPEG})
 set(HAVE_MMAL ${WITH_MMAL})
@@ -37,30 +37,28 @@ set(HAVE_MYSQL ${WITH_MYSQL})
 set(HAVE_PGSQL ${WITH_PGSQL})
 set(HAVE_SQLITE3 ${WITH_SQLITE3})
 
-set(WITHOUT_V4L2 ON)
-  if(CMAKE_SYSTEM_NAME MATCHES "FreeBSD" AND NOT(WITH_PWCBSD))
+set(HAVE_V4L2 OFF)
+  check_include_files("linux/videodev2.h" HAVE_LINUX_VIDEODEV2_H)
+  if(HAVE_LINUX_VIDEODEV2_H AND WITH_V4L2)
+    set(HAVE_V4L2 ON )
+  endif(HAVE_LINUX_VIDEODEV2_H AND WITH_V4L2)
+
+set(HAVE_BKTR OFF)
+  if(CMAKE_SYSTEM_NAME MATCHES "FreeBSD")
     check_include_files("dev/bktr/ioctl_bt848.h" HAVE_FREEBSD_BT848)
     check_include_files("dev/bktr/ioctl_meteor.h" HAVE_FREEBSD_METEOR)
-    if(HAVE_FREEBSD_BT848 AND HAVE_FREEBSD_METEOR AND WITH_V4L2 )
-      set(WITHOUT_V4L2 OFF )
-    endif(HAVE_FREEBSD_BT848 AND HAVE_FREEBSD_METEOR AND WITH_V4L2 )
-  endif(CMAKE_SYSTEM_NAME MATCHES "FreeBSD" AND NOT(WITH_PWCBSD))
+    if(HAVE_FREEBSD_BT848 AND HAVE_FREEBSD_METEOR AND WITH_BKTR)
+      set(HAVE_BKTR ON )
+    endif(HAVE_FREEBSD_BT848 AND HAVE_FREEBSD_METEOR AND WITH_BKTR)
+  endif(CMAKE_SYSTEM_NAME MATCHES "FreeBSD")
 
   if(CMAKE_SYSTEM_NAME MATCHES "NetBSD" OR CMAKE_SYSTEM_NAME MATCHES "OpenBSD")
     check_include_files("dev/ic/bt8xx.h" HAVE_NETOPEN_BT8XX)
-    if(HAVE_NETOPEN_BT8XX AND WITH_V4L2)
-      set(WITHOUT_V4L2 OFF )
-    endif(HAVE_NETOPEN_BT8XX AND WITH_V4L2)
+    if(HAVE_NETOPEN_BT8XX AND WITH_BKTR)
+      set(HAVE_BKTR ON )
+    endif(HAVE_NETOPEN_BT8XX AND WITH_BKTR)
   endif(CMAKE_SYSTEM_NAME MATCHES "NetBSD" OR CMAKE_SYSTEM_NAME MATCHES "OpenBSD")
 
-  if(CMAKE_SYSTEM_NAME MATCHES "Linux")
-    check_include_files("linux/videodev2.h" HAVE_LINUX_VIDEODEV2_H)
-    if(HAVE_LINUX_VIDEODEV2_H AND WITH_V4L2)
-      set(WITHOUT_V4L2 OFF )
-    endif(HAVE_LINUX_VIDEODEV2_H AND WITH_V4L2)
-  endif(CMAKE_SYSTEM_NAME MATCHES "Linux")
-
-set(PWCBSD WITH_PWCBSD)
 
 configure_file(config.h.in "${CMAKE_CURRENT_SOURCE_DIR}/config.h")
 configure_file(motion-dist.conf.in motion-dist.conf)
@@ -77,13 +75,8 @@ configure_file(motion.init-FreeBSD.sh.in motion.init-FreeBSD.sh)
 list(APPEND SRC_FILES
      conf.c motion.c alg.c draw.c event.c ffmpeg.c jpegutils.c logger.c md5.c
      netcam.c netcam_ftp.c netcam_jpeg.c netcam_rtsp.c netcam_wget.c
-     picture.c rotate.c stream.c track.c vloopback_motion2.c webhttpd.c)
-if(CMAKE_SYSTEM_NAME MATCHES "FreeBSD" AND NOT(WITH_PWCBSD))
-	list(APPEND SRC_FILES video_freebsd.c)
-else()
-	list(APPEND SRC_FILES video2.c video_common.c)
-endif()
-
+     picture.c rotate.c stream.c track.c video_loopback.c webhttpd.c
+     video_v4l2.c video_common.c video_bktr.c)
 include_directories(${JPEG_INCLUDE_DIR})
 list(APPEND LINK_LIBRARIES ${CMAKE_THREAD_LIBS_INIT} ${JPEG_LIBRARIES})
 
@@ -122,12 +115,13 @@ message("--   Config Dir:             " ${sysconfdir})
 message("--   Bin Dir:                " ${BIN_PATH})
 message("--   JPEG Include Dir:       " ${JPEG_INCLUDE_DIR})
 message("--   WITH_V4L2:              " ${WITH_V4L2})
-message("--   Without V4L2:           " ${WITHOUT_V4L2})
-message("--   With PWC BSD:           " ${WITH_PWCBSD})
-message("--   MYSQL:   " ${WITH_MYSQL} " Include Dir: " ${MYSQL_INCLUDE_DIRS} )
+message("--   HAVE_V4L2:              " ${HAVE_V4L2})
+message("--   WITH_BKTR:              " ${WITH_BKTR})
+message("--   HAVE_BKTR:              " ${HAVE_BKTR})
+message("--   MYSQL:   " ${WITH_MYSQL}   " Include Dir: " ${MYSQL_INCLUDE_DIRS} )
 message("--   SQLITE3: " ${WITH_SQLITE3} " Include Dir: " ${SQLITE3_INCLUDE_DIRS} )
-message("--   PGSQL:   " ${WITH_PGSQL} " Include Dir: " ${PostgreSQL_INCLUDE_DIRS} )
-message("--   FFMPEG:  " ${WITH_FFMPEG} " Include Dir: " ${FFMPEG_INCLUDE_DIRS} )
+message("--   PGSQL:   " ${WITH_PGSQL}   " Include Dir: " ${PostgreSQL_INCLUDE_DIRS} )
+message("--   FFMPEG:  " ${WITH_FFMPEG}  " Include Dir: " ${FFMPEG_INCLUDE_DIRS} )
 message("--   MMAL:    " ${WITH_MMAL} )
 message("--   CFLAGS:  " ${CMAKE_C_FLAGS} )
