+++ motion.1
@@ -165,8 +165,8 @@ Description:
 .fi
 .RS
 String to specify the videodevice to be used for capturing.
-For Linux, the format is usually /dev/videoX where X varies depending upon the video devices connected to the computer.
-For FreeBSD the default device would be /dev/bktr0.
+The format is usually /dev/videoX where X varies depending upon the video devices connected to the computer.
+For FreeBSD certain devices use the bktr subsystem and they will use /dev/bktr0.
 .RE
 .RE
