+++ motion.h
@@ -47,6 +47,10 @@
 #include <stdint.h>
 #include <pthread.h>
 
+#if defined(BSD)
+#include <pthread_np.h>
+#endif
+
 #include "logger.h"
 #include "conf.h"
 #include "stream.h"
@@ -132,6 +136,9 @@
 #define DEF_FFMPEG_BPS      400000
 #define DEF_FFMPEG_VBR           0
 #define DEF_FFMPEG_CODEC   "mpeg4"
+
+#define DEF_INPUT               -1
+#define DEF_VIDEO_DEVICE         "/dev/video0"
 
 #define THRESHOLD_TUNE_LENGTH  256
 
@@ -226,6 +233,14 @@ struct images;
 #define IMAGE_SAVED      8
 #define IMAGE_PRECAP    16
 #define IMAGE_POSTCAP   32
+
+enum CAMERA_TYPE {
+    CAMERA_TYPE_UNKNOWN,
+    CAMERA_TYPE_V4L2,
+    CAMERA_TYPE_BKTR,
+    CAMERA_TYPE_MMAL,
+    CAMERA_TYPE_NETCAM
+};
 
 struct image_data {
     unsigned char *image;
@@ -347,10 +362,13 @@ struct context {
     struct config conf;
     struct images imgs;
     struct trackoptions track;
+
+    enum CAMERA_TYPE      camera_type;
     struct netcam_context *netcam;
 #ifdef HAVE_MMAL
     struct mmalcam_context *mmalcam;
 #endif
+
     struct image_data *current_image;        /* Pointer to a structure where the image, diffs etc is stored */
     unsigned int new_img;
