+++ track.c
@@ -8,7 +8,7 @@
 #include <math.h>
 #include "motion.h"
 
-#ifndef WITHOUT_V4L2
+#ifdef HAVE_V4L2
 #include <linux/videodev2.h>
 #include "pwc-ioctl.h"
 #endif
@@ -52,14 +52,14 @@ static unsigned int servo_move(struct context *cnt, struct coord *cent,
                                      struct images *imgs, unsigned int manual);
 static unsigned int iomojo_move(struct context *cnt, int dev, struct coord *cent, struct images *imgs);
 
-#ifndef WITHOUT_V4L2
+#ifdef HAVE_V4L2
 static unsigned int lqos_center(struct context *cnt, int dev, int xoff, int yoff);
 static unsigned int lqos_move(struct context *cnt, int dev, struct coord *cent,
                                     struct images *imgs, unsigned int manual);
 static unsigned int uvc_center(struct context *cnt, int dev, int xoff, int yoff);
 static unsigned int uvc_move(struct context *cnt, int dev, struct coord *cent,
                                    struct images *imgs, unsigned int manual);
-#endif /* WITHOUT_V4L2 */
+#endif /* HAVE_V4L2 */
 
 /* Add a call to your functions here: */
 unsigned int track_center(struct context *cnt, int dev ATTRIBUTE_UNUSED,
@@ -79,7 +79,7 @@ unsigned int track_center(struct context *cnt, int dev ATTRIBUTE_UNUSED,
     } else if (cnt->track.type == TRACK_TYPE_SERVO) {
         return servo_center(cnt, xoff, yoff);
     }
-#ifndef WITHOUT_V4L2
+#ifdef HAVE_V4L2
     else if (cnt->track.type == TRACK_TYPE_PWC)
         return lqos_center(cnt, dev, xoff, yoff);
     else if (cnt->track.type == TRACK_TYPE_UVC)
@@ -108,7 +108,7 @@ unsigned int track_move(struct context *cnt, int dev, struct coord *cent, struct
         return stepper_move(cnt, cent, imgs);
     else if (cnt->track.type == TRACK_TYPE_SERVO)
         return servo_move(cnt, cent, imgs, manual);
-#ifndef WITHOUT_V4L2
+#ifdef HAVE_V4L2
     else if (cnt->track.type == TRACK_TYPE_PWC)
         return lqos_move(cnt, dev, cent, imgs, manual);
     else if (cnt->track.type == TRACK_TYPE_UVC)
@@ -780,7 +780,7 @@ static unsigned int iomojo_move(struct context *cnt, int dev, struct coord *cent
     Logitech QuickCam Orbit camera tracking code by folkert@vanheusden.com
 
 ******************************************************************************/
-#ifndef WITHOUT_V4L2
+#ifdef HAVE_V4L2
 static unsigned int lqos_center(struct context *cnt, int dev, int x_angle, int y_angle)
 {
     int reset = 3;
@@ -927,7 +927,15 @@ static unsigned int uvc_center(struct context *cnt, int dev, int x_angle, int y_
         int reset = 3; //0-non reset, 1-reset pan, 2-reset tilt, 3-reset pan&tilt
         struct v4l2_control control_s;
 
-        control_s.id = V4L2_CID_PANTILT_RESET;
+        control_s.id = V4L2_CID_PAN_RESET;
+        control_s.value = (unsigned char) reset;
+
+        if (ioctl(dev, VIDIOC_S_CTRL, &control_s) < 0) {
+            MOTION_LOG(ERR, TYPE_TRACK, SHOW_ERRNO, "%s: Failed to reset UVC camera to starting position! Reason");
+            return 0;
+        }
+
+        control_s.id = V4L2_CID_TILT_RESET;
         control_s.value = (unsigned char) reset;
 
         if (ioctl(dev, VIDIOC_S_CTRL, &control_s) < 0) {
@@ -1071,7 +1079,15 @@ static unsigned int uvc_move(struct context *cnt, int dev, struct coord *cent,
         unsigned int reset = 3; //0-non reset, 1-reset pan, 2-reset tilt, 3-reset pan&tilt
         struct v4l2_control control_s;
 
-        control_s.id = V4L2_CID_PANTILT_RESET;
+        control_s.id = V4L2_CID_PAN_RESET;
+        control_s.value = (unsigned char) reset;
+
+        if (ioctl(dev, VIDIOC_S_CTRL, &control_s) < 0) {
+            MOTION_LOG(ERR, TYPE_TRACK, SHOW_ERRNO, "%s: Failed to reset UVC camera to starting position! Reason");
+            return 0;
+        }
+
+        control_s.id = V4L2_CID_TILT_RESET;
         control_s.value = (unsigned char) reset;
 
         if (ioctl(dev, VIDIOC_S_CTRL, &control_s) < 0) {
@@ -1219,4 +1235,4 @@ static unsigned int uvc_move(struct context *cnt, int dev, struct coord *cent,
 
     return cnt->track.move_wait;
 }
-#endif /* WITHOUT_V4L2 */
+#endif /* HAVE_V4L2 */
