+++ video_loopback.c
@@ -1,5 +1,5 @@
 /*
- *    vloopback_motion.c
+ *    video_loopback.c
  *
  *    Video loopback functions for motion.
  *    Copyright 2000 by Jeroen Vreeken (pe1rxq@amsat.org)
@@ -8,17 +8,16 @@
  *    See also the file 'COPYING'.
  *
  */
-#include "vloopback_motion2.h"
-#if (!defined(WITHOUT_V4L2)) && (!defined(BSD))
+#include "motion.h"
+
+#if (defined(HAVE_V4L2)) && (!defined(BSD))
+
+#include "video_loopback.h"
 #include <dirent.h>
 #include <sys/ioctl.h>
 #include <linux/videodev2.h>
 
-/**
- * v4l2_open_vidpipe
- *
- */
-static int v4l2_open_vidpipe(void)
+static int vlp_open_vidpipe(void)
 {
     int pipe_fd = -1;
     char pipepath[255];
@@ -105,7 +104,7 @@ typedef struct capent {const char *cap; int code;} capentT;
         {"Last",0}
 };
 
-static void show_vcap(struct v4l2_capability *cap) {
+static void vlp_show_vcap(struct v4l2_capability *cap) {
     unsigned int vers = cap->version;
     unsigned int c    = cap->capabilities;
     int i;
@@ -122,7 +121,7 @@ static void show_vcap(struct v4l2_capability *cap) {
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: ------------------------");
 }
 
-static void show_vfmt(struct v4l2_format *v) {
+static void vlp_show_vfmt(struct v4l2_format *v) {
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: type: type:           %d",v->type);
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: fmt.pix.width:        %d",v->fmt.pix.width);
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: fmt.pix.height:       %d",v->fmt.pix.height);
@@ -134,18 +133,14 @@ static void show_vfmt(struct v4l2_format *v) {
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: ------------------------");
 }
 
-/**
- * v4l2_startpipe
- *
- */
-static int v4l2_startpipe(const char *dev_name, int width, int height, int type)
+int vlp_startpipe(const char *dev_name, int width, int height)
 {
     int dev;
     struct v4l2_format v;
     struct v4l2_capability vc;
 
     if (!strcmp(dev_name, "-")) {
-        dev = v4l2_open_vidpipe();
+        dev = vlp_open_vidpipe();
     } else {
         dev = open(dev_name, O_RDWR);
         MOTION_LOG(NTC, TYPE_VIDEO, NO_ERRNO, "%s: Opened %s as pipe output", dev_name);
@@ -161,7 +156,7 @@ static int v4l2_startpipe(const char *dev_name, int width, int height, int type)
         return -1;
     }
 
-    show_vcap(&vc);
+    vlp_show_vcap(&vc);
 
     memset(&v, 0, sizeof(v));
 
@@ -172,18 +167,18 @@ static int v4l2_startpipe(const char *dev_name, int width, int height, int type)
         return -1;
     }
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: Original pipe specifications");
-    show_vfmt(&v);
+    vlp_show_vfmt(&v);
 
     v.type = V4L2_BUF_TYPE_VIDEO_OUTPUT;
     v.fmt.pix.width = width;
     v.fmt.pix.height = height;
-    v.fmt.pix.pixelformat = type;
+    v.fmt.pix.pixelformat = V4L2_PIX_FMT_YUV420;
     v.fmt.pix.sizeimage = 3 * width * height / 2;
     v.fmt.pix.bytesperline = width;
     v.fmt.pix.field = V4L2_FIELD_NONE;
     v.fmt.pix.colorspace = V4L2_COLORSPACE_SRGB;
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: Proposed pipe specifications");
-    show_vfmt(&v);
+    vlp_show_vfmt(&v);
 
     if (ioctl(dev,VIDIOC_S_FMT, &v) == -1) {
         MOTION_LOG(ERR, TYPE_VIDEO, SHOW_ERRNO, "%s: ioctl (VIDIOC_S_FMT)");
@@ -191,35 +186,15 @@ static int v4l2_startpipe(const char *dev_name, int width, int height, int type)
     }
 
     MOTION_LOG(INF, TYPE_VIDEO, NO_ERRNO, "%s: Final pipe specifications");
-    show_vfmt(&v);
+    vlp_show_vfmt(&v);
 
     return dev;
 }
 
-/**
- * v4l2_putpipe
- *
- */
-static int v4l2_putpipe(int dev, unsigned char *image, int size)
+int vlp_putpipe(int dev, unsigned char *image, int imgsize)
 {
-    return write(dev, image, size);
+    return write(dev, image, imgsize);
 }
 
-/**
- * vid_startpipe
- *
- */
-int vid_startpipe(const char *dev_name, int width, int height, int type)
-{
-    return v4l2_startpipe(dev_name, width, height, type);
-}
 
-/**
- * vid_putpipe
- *
- */
-int vid_putpipe (int dev, unsigned char *image, int size)
-{
-    return v4l2_putpipe(dev, image, size);
-}
-#endif /* !WITHOUT_V4L2 && !BSD */
+#endif /* HAVE_V4L2 && !BSD */
